refactor LocalCachedSearchDataSource to have the needed functionality

+++ data/local_data_source/search/src/main/java/com/sanaa/search/data/local/dao/SearchDao.kt
@@ -4,13 +4,13 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.sanaa.search.data.local.entity.SearchLocalDto
+import com.sanaa.search.dataSource.local.dto.SearchLocalDto
 
 @Dao
 interface SearchDao {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertSearch(search: SearchLocalDto): Long
+    suspend fun insertSearch(search: SearchLocalDto): Int
 
     @Query("SELECT * FROM searches WHERE `query` = :query AND language = :language LIMIT 1")
     suspend fun getSearchByQueryAndLanguage(query: String, language: String): SearchLocalDto?
